third commit

+++ src/main/java/com/example/model/NoteNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.model;
 
 public class NoteNotFoundException extends RuntimeException {
-    public NoteNotFoundException(int id) {
+    public NoteNotFoundException(long id) {
         super("Note with id " + id + " not found");
     }
 }
